Create README.md
***Project Description***:
Spaceman is a space-themed adaptation of the classic web-based hangman game. It's a widely enjoyed puzzle game where a ruthless computer conceals a word from a player, and the player tries to guess letter by letter. If the player fails, a hazardous spaceship will be constructed and may harm the planet. If the player wins, humanity survive.

***Wire Frames***: 
Copy and paste or drag and drop your images here. Wireframe

***User Stories***: 
As a player, I want to select the game's level. As a player, I want to know my letter guesses, which ones are correct and in the right position, and which ones are incorrect. As a player, I want to have a limited number of guessing attempts based on the chosen difficulty level. As a player, I want a clear indication of my game progress, such as how many letters of the secret word I've guessed correctly and how many attempts I have left. As a player, I want to know when I have won/lost the game.

***MVP Goals***:
Allow player selects the difficulty level. Generate a random word for the player to guess. Allow the player to click on the letters (buttons) to guess the hidden word. Allow limited number of guessing attempts, with the limit determined by the chosen difficulty level. Display a spaceship construction progress when incorrect guesses are made.

**Level Easy**: A guessing word consists around 5-7 letters. A player have more guessing attempts.

**Level Insane**: A guessing word consists more than 10 letters. A player have less guessing attempts. A timer will be displayed.

***Stretch Goals***:
As a developer, I would like to build different levels of difficulty. I would like to add the timer to specific level to increase tension and urgency. I would like to have the choice to turn music on or off while it plays in the background. I would like to offer a player the option to use hints to make the game more accessible. I would like to enhance the game's code and experience, making Spaceman more engaging, easy to understand and enjoyable.